BUILD/MAJOR: Prebuild now includes SWIG auto regenerate check
When the C++ Trie or Pattern DLL is rebuilt the pre-build event is used to determine if SWIG is present on the system. If SWIG is present then the auto generated C++ and C# files are regenerated from the 51Degrees.i file in src. If not then the source included in the repository is used. In all cases a message is included in the build output displayed in Visual Studio.