BUG/MINOR: Copies properties before altering them
The original implementation would corrupt the string array passed to the method which could produce unexpected consequences for the caller. The change makes a copy of the string before replacing separators with NULL string terminators.

Additionally const memory is now cast to void* before freeing.

The change will only impact those that reuse the string passed into the method and found this defect.