REORG/MINOR: Uses new V3.2 wrappers
Wrapper matcher classes are changed to use updated wrappers. This only
affects the internal interface with the wrapper modules.

Input and output is unchanged, so shouldn't affect usage. Only
'v3-wrapper' and 'v3-trie-wrapper' are supported, so if you use an
outdated method name (e.g. 'trie-wrapper) in you settings file, an error
will occur.

If a property is requested which is not available in your data file, a
null will be returned as was the case previously.