FEAT/MAJOR: implemented support for Trie wrapper
The 51Degrees C extension for PHP has initially only supported Pattern
matching. Trie matching has now been implemented to allow users to
access our fastest matching algorithm.

# PHP functions
Trie implements the same set of functions exposed to PHP as Pattern.
Switching from one to another will not require any changes to the
existing logic.

Extension functions accessible from PHP are:
- fiftyone_info: returns dataset initialisation status with a short
message briefly describing the initialisation result and the data file
type as Trie.
- fiftyone_get_http_headers: returns an array of all HTTP headers that
the detector can use. Values are not returned and should be retrieved
from the _SERVER[] array if required.
- fiftyone_match: performs an automatic match based on all available
HTTP headers. No input is required as all of the relevant HTTP headers
are retrieved from the _SERVER array automatically. Returns a PHP array
with Property => Value pairs.
- fiftyone_match_with_useragent: performs a match based on the user-
supplied User-Agent HTTP header. The User-Agent HTTP header is normally
retrieved from the _SERVER array, but other sources can be used as well.
Function returns a PHP array of Property => Value pairs.
- fiftyone_degrees_get_properties: function is identical to
fiftyone_match_with_useragent: requires a User-Agent HTTP header input
and returns a PHP array of Property => Value pairs. Exists for
compatibility purposes with previous implementations. Will be removed in
future versions.
- fiftyone_match_with_headers: performs a match based on user-supplied
string of HTTP headers and values. The header must be separated from the
value by a white space or colon. Each header value pair must end with a
new line symbol '\n'. Returns a PHP array with Property => Value pairs.

#Internal Functions
Trie implements several private helper functions that process user input
and return result.
- buildArray: constructs a PHP array with detection results. Array to
store results is allocated in the calling function and passed to this
function.
- clearOffsets: resets the count of device offsets to 0, allowing to
reuse the allocated memory.
- matchHttpHeaders: function takes in the user-supplied string with HTTP
headers and values and populates device offsets with the relevant
headers and values. Invokes buildArray.
- matchUserAgent: takes in a single string containing the HTTP User-
Agent header and creates a single device offset. Invokes the buildArray
function.
- datasetInitSuccess: returns 1 or 0 depending on whether the dataset
initialisation was successful. Used in functions exposed to the PHP to
stop execution if there was a problem with dataset initialisation.

#Initialisation and Shutdown
The data file is read into memory upon module initialisation. Location
of the detector and the data file is supplied by the relevant PHP.ini
settings.
Upon module shutdown the memory allocated for the dataset is released.
A small piece of memory is allocated for each request at request
initialisation. This memory is used to store deviceOffsets used for
device detection. The amount of memory allocated will always be equal to
the number of important headers times the size of the deviceOffset. At
the end of each request memory is freed.

#Building the extension.
Extension can be built using the provided config.m4 file and consists of
three steps:
1. Navigate to the Device-Detection/php/trie folder and run phpize5
2. run ./configure
3. run sudo make install

The Trie implementation of the PHP C extension has been tested by
51Degrees on at least two independent environments and results were
found to be consistent on both. The examples.php page provides a
demonstration and comparison of the output of each match function
exposed to PHP. Firefox addon called 'Modify Headers' was used to test
various situations where different combinations of HTTP headers were
presented to the detector.

Please note that Trie requires a very large continuous chunk of RAM to
be allocated to read the data file into. If detector fails to initialise
please ensure that you have at lest 1.5Gb of free RAM. In addition the
32-bit systems normally do not support allocating chunks of RAM of this
size so Premium and Enterprise data file initialisation is expected to
fail on a 32-bit system.