FEAT/MAJOR: Added support for SWIG and C++
Now includes C++ classes for both Pattern and Trie algorithms. The classes are interface compatible with the exception of some constructors. They can be used interchangeably once created. The Pattern Provider handles caching and pooling but is otherwise identical to the same classes in Trie.

SWIG is used to create wrappers initially .NET, Python and Perl. Other supported high level languages could be used.

This code is new and as such should be used with caution when upgrading from prior versions. It has been tested with the full set of C/C++ pre-deployment tests.