EXAMPLE/MEDIUM: Created example PHP page
Previously the 51Degrees PHP C extension was provided without any
examples that could simply be copied into the Web directory and accessed
via a Web browser.

Since the functionality of the API has expanded we decided to provide an
example that would illustrate how to use the API in a nice and
consistent way throughout all of our C based packages, including PHP.

The large chunk of the example PHP source code is used for presenting
the data in an easy to read table with links and a list of relevant HTTP
headers for each function.

Example page will work with both Pattern and Trie versions of the
detector.