BUG/MINOR: Address fault to numeric match and max 16 bit integers
The if statement in setNumericNodeState checks for less than or equal to, but the setRange method was not considered a value of SHRT_MAX (32767). If the state->target value was 32767 then the if statement in the caller would allow the numeric match to be tested, but the setRange method would not set a range to be checked. As a result the state->range pointer would be unset and the memory error occurred. The change addresses this rare situation.

Renamed fiftyoneDegreesRange structure to follow standard naming convention.