FEAT: Added support to reload the data set from file or memory. Implemented a provider model to encapsulate data sets, cache and pool.
A new structure called fiftyoneDegreesProvider is used to link to an active pool and associated locks. The provider is used to gain access to work sets and is supported by the new reload methods. When used in a multi threaded environment (the default unless FIFTYONEDEGREES_NO_THREADING is specified) the associated methods are thread safe in GCC and MSVC compilers.

When the reload method is used there is a brief period where the amount of required memory doubles so that the new and old data sets can coexist. This should be considered when planning memory management.

The following new methods have been added to support the provider structure and reloading the associated data set.

fiftyoneDegreesInitProviderWithPropertyArray – initialises a new provider from a file and an array of properties that the associated data set should return.

fiftyoneDegreesInitProviderWithPropertyString – initialises a new provider from a file and a character separated list of properties that the associated data set should return.

fiftyoneDegreesProviderFree – releases all the resources used by the provider.

fiftyoneDegreesProviderWorksetGet – the only way to safely return a work set to use for device detection from the provider. This method is thread safe.

fiftyoneDegreesWorksetRelease – returns the work set provided back to the pool and provider that it came from. If the pool has been replaced the last work set returned will trigger the release of associated resources. This method is thread safe.

fiftyoneDegreesProviderReloadFromFile – creates a new data set, pool and cache from the same file location used to create the original provider. The existing resources are released when the last work set is returned, or if already free are released by the method.

fiftyoneDegreesProviderReloadFromMemory – like reload from file but uses a pointer to provide the data to be used with the new data set.

Deprecated Methods

fiftyoneDegreesWorksetPoolRelease should no longer be used as the work set now stores a reference to the pool that it came from. As such there is no longer a need for a method that takes both a pool and a work set pointer.

Work sets

All the work sets that can be contained in a pool are now created and assigned to the pool at the time the pool is created. This modification reduces the complexity associated with managing pools and results in consistent memory usage.

Work sets now store a reference to the pool that they are associated with. If the work set is not associated with a pool then this reference is NULL. The reference is used to return the work set to the pool it came from as there is now the possibility that several pools can be in existence simultaneously during a reload operation in a multi-threaded implementation.

Threading

To support the change locking and signalling methods needed to be modified so they could be references as pointers by MSVC and GCC. As such there have been extensive changes to the threading.h source code. Threading.c has been added to provide more complex implementations for GCC compilers.

Find Profiles

The methods associated with finding profiles have been moved into their own region. Some of the methods have been modified slightly to address minor issues when used on different compilers.

fiftyoneDegrees Prefixed Methods

Methods prefixed fiftyoneDegrees are intended to be exposed to external source code. However in previous versions some methods were marked with the prefix when they were not exposed in this way. These methods have had the prefix removed and are now marked static so that they can only be used from within 51Degrees.c.

Static Methods

All methods not used outside 51Degrees.c are now marked with static to reduce any possibility of conflict with third party source code and libraries.

Code Comments & Style

Some code comments have been improved with a particular emphasis on explaining the locking strategy used to synchronise access to resources.

C++ Classes & SWIG

The associated C++ classes have now been modified to use the provider rather than a pool directly.

Provider class now contains the method reloadFromFile which maps to the C fiftyoneDegreesProviderReloadFromFile method.

.NET Interop

The interoperability layer between managed and unmanaged code has been modified to expose the new functionality. Classes related to SWIG that are included in git have been modified automatically.