BUILD/EXAMPLE/OPTIM: Moved node module to pattern directory.
Created examples directory with all the common examples, they do not yet
have the copyright, snippet or tutorial explanation.
in FiftyOneDegreesPattern.js, an availableProperties array has been
added so they are accessible in a normal node way rather that using
get() and not being able to print the full contents of the array.

Related Work Items: #703 #700