BUILD/CLEANUP: Added a trie directory which builds in the same way as pattern.
In doing so, also made the module common and put into the root node
directory, this will then include either pattern or trie depending on
the data file it is given. This means the FiftyOneDegrees**V3.js files
have been replaced by a single node.js/FiftyOneDegreesV3.js and the
update.js has been moved to node.js/.
Some changes to examples, config, binding and test have been made to
accomodate the new directory structure.
In src/trie/ a dispose method has been added to the match class in
Match.* and 51Degrees.i. This is identical to the virtual dispose method
but is declared confitionaly for node using correct nomenclature.
In src/pattern the provider has been cleaned up, removing the unused
destroy function that was created.

Related Work Items: #695