TEST/REORG: Tests added and existing tests restructured.
API function tests have been added for find profiles and match for
device id. The find profiles test uncovered a bug which has been fixed
(see #1139).
The tests for single User-Agent matching have been changed so that the
assert statements for a match existing and the match being correct are
in their own "it" functions, meaning they are displayed separately when
the tests are run. This may not be possible in the same way for the
tests that use the User-Agents file, but it should be done, probably by
adding a "passed" variable which is checked at the end. The reason for
this is that doing it in the same way means the passed  message is
output for every User-Agent, so 20000 times for each test.
The cache test has been temporarily disabled as there is a bug meaning
it fails, this has a todo.
Also, the node SWIG output has been added. This is the product of the
ealier commit fixing #1139.

Related Work Items: #696 #1139