BUG: When the data file is not found, the program tries to free things that arent there.
Now in 51Degrees.c dataSet is freed if the init was unsuccessful rather
than doing a full datasetfree on provider->active->dataSet which would
be null, and the just the dataset structure and not its elements are
freed.
In Provider.cpp fiftyonedegreesproviderfree is no longer called, as if
the provider has failed to initialise, everything will have been freed
already.

Related Work Items: #1161