BUG: Fixed the bug where a specific error was not thrown, just a general swig one.
The problem was that because the function is an overloaded one, the
correct function is called, then either returns the provider or
continues if there was an error thrown. This continuing means that it
just runs into the default error and returns that instead. So overriding
swigs template and adding a goto fail means when there is an error
thrown it skips over the default error and ends the outer function.

Related Work Items: #1140