FEAT/OPTIM: Added usage sharing and removed unneccessary code.
Usage sharing has been added in shareUsage.js. This is initialised from
FiftyOneDegreesV3.js id it is enabled in the config. It will need to be
changed so it is enabled unless otherwise specified. The usage sharing
uses the same structure and function names as the Java webapp and has
been tested sending data to a locally listening port. Key differences
from the Java API are:
- this is not a background thread as node cannot do that, it could be
changed to use "nextTick" which would thread share but would probably
still have the same performace impact.
- The ips are stripped of the first part (prefixing ':') before they are
sent.
- Error codes are different but are dealt with in a similar way.
- The XML is not sent as a stream, the whole lot is appended to a string
the sent.

In FiftyOneDegreesV3 the function 'getMatchForRequest' has been added to
take an http request and return a match, this is where usage sharing is
implemented (client.js has been changed to use this).
Where the provider is initialised, the ifs have been optimised. Where a
check for trie was done to rule out cacheSize and poolSize, this has
been removed as these will just be skipped anyway. A check for whether
they exist may be needed when trie is used though, but in a cleaner
looking way.
Temporarily the logging has been set to print any errors to the console.

Related Work Items: #1020 #1081 #1082 #1083