FEAT: Logging has been added and comments improved.
In FiftyOneDegreesV3, proper logging has been added. There are now three
levels of logging: error, info, and debug. This is defined in the config
as logLevel and has a default of info if not set.
Errors are no longer thrown when the provider cannot be initialised,
they are logged and null returned. This means a program wont break, but
the honus is on the developer to check for an null returned provider.

There is now much more error catching in update, response codes and
licance key are checked. Any errors in update are logged as info.
The 'Product' part of the query string has been removed as it is not
used in the java implementation, the 'binaryv3' and 'trie' is enough.

In all files there has been a general tidy up and more explanation as to
what is happening added in the comments.

Related Work Items: #1020 #1083 #1129