DOC/EXAMPLE: Example explanations have been added and logging removed.
In all examples,  full exmplanations have been added in line with other
languages. The layout of the code and some comments have also been
changed in line with other languages.
In the tests, the config has been aligned properly and the text for the
find non-mobile profiles test corrected.
In both the examples and tests, the log level has been set to none so
the is no unnecessary text ouput.

Related Work Items: #700 #696