EXMAPLE: Client and clientTrie now one file and express example added.
clientTrie has been removed and the option added in client to use trie
by adding trie as an argument. If there are no arguments it defaults to
pattern.
expressClient has been added, this uses express which is an extension of
the http module. Because it has the get function which takes the url as
an argument, a metrics subpage was added as being seperate the code is
still simple and easy to read.
Both now print all values by using the getValues function instead of the
getValue, and print which method is being used and which port on
startup.

Related Work Items: #1047