FEAT/EXAMPLE: The config option nodeify has been added and used in examples.
When returning a match using either the matchForHttpHeaders or
matchForRequest functions with the nodeify option set to true, all the
properties are set as keys in the match object. The getters of these are
such that properties with multiple values are returned as arrays so can
be easily printed without using a for loop, and boolean values are
retured as a boolean so doing things like if(match.IsMobile) is
possible. The values have been assigned getters rather than all put into
an array so that only the properties that are needed are actually
fetched from the C API.

This new feature has been used in the client and expressClient examples
meaning there are less lines of code and it is done in a slightly more
node way.

The code in general in both examples has been tidied by running
beautify.io, and match renamed to device.

Related Work Items: #1047 #1188