FEAT: Edited the getMatchForRequest function to automatically dispose and define getters.
When provider.getMatchForRequest is used, it no longer returns a match
object, it will attach the match object to the request which can be used
as normal via request.match. All the available properties, with the
exception of JavasciptHardwareProfile, have getters assigned to the
request object which will fetch the value from the match object. The
properties with multiple values will return an array, and the properties
with a single value will just return that. In addition, the boolean
values are returned as a boolean rather than a string.
Examples:
request.IsMobile returns bool
request.BrowserName returns string
request.SupportedBearers returns array.

This exra functionality can be disabled by setting "addGetters : false"
in the config file. This will save some compute time, but it doesn't
seem to make much difference.

The function also now attaches the match dispose method to request.end,
request.abort and request.aborted. This means that the match object does
not need to be manually disposed of when using the getMatchForRequest
method.

In the two web examples, this new method has been added to demonstrate
the functionality.

Note: usage sharing will only happen when using getMatchForRequest as this is the only method which will be passed an ip.

Related Work Items: #697 #1047