FEAT/BUILD: Added generic match and getters, and added package.json.
in FiftyOneDegreesV3 there is now a match function which directs
whatever input to the correct matching function. It also calls a new
defineGetter function which sets the getters for each property meaning
match.IsMobile can be used in place of match.getValue("IsMobile"), and
more importantly, match.SupportedBearers can be used in place of...
vals = match.getValues("SupportedBearers")
arr = new Array(vals.size())
for (i = 0; i < vals.size(); i++) {
arr[i] = vals.get(i)
}
which is much neater.
All the examples have changed to use this new method, and new tests have
been added for it. The current tests have stayed the same so the core
functions and the wrapper are tested separately.

a package.json file has been included, and the binding.gyp's
consolidated into one. So the module can now be installed with "npm
install". This change requires that the src folder exists in the node.js
directory, so this must be added as an automated step when building from
git.

".dat" has been removed from .npmignore so a lite file is now included.
A step must be added to copy the data file from the data directory to
the node.js/data directory, and the path set as the default data file if
none is supplied.

Related Work Items: #1021 #1202 #1207