BUILD: Copies source and data when building from git.
When using git, the source (and data) will be in a directory up one
level, so it will now be copied to a directory that npm expects. This is
done in gitCopy.js which is called as a preinstall in package.json.
The src and data directories have been added to gitignore so they are
removed from git, but not npm.

In addition, msvs version has been added to package.json so it will
compile in windows.

Related Work Items: #1213