CLEANUP: Improved error handling.
Events are now emitted without the '51' prefix. This means that if debug
level is set to none and 'error' is emitted, a full on error is thrown.
There is an extra listener which listens for 'error' and will log the
error to a console if there are no listeners instead of throwing the
error.

Related Work Items: #1203