BUG: Fixed getProviderSize in GCC and threading problem in perl.
the getProviderSize implementation in Provider.cpp has been altered to
allow compilation in GCC. This has meant the inclusion of stdlib.h in
Provider.hpp and a change to the call to the constuctor when validate =
false.

In the Makefile.PL threading has been included in the linker so there is
not a runtime error when trying to use a mutex.

Related Work Items: #1385