FEAT/TEST/BUG: Added getProviderSize and tests to Trie and fixed a couple of things.
getProviderSize and associated c++ constructor and c# tests have now been added to Trie in the same way as in Pattern. The full explanation to this is in the previous commit.

in the getProviderSize function the size of the provider structure has been added, and a bug uncovered and fixed. When looking at the properties string, the getSeparatorCount function ids used. This means that one property is dropped, so it has been added.

All API tests that are used for Pattern are now used for Trie as well. This means there is also a new TrieBase class in the unit tests. To run the API tests properly, the TrieWrapper has been refactored. The previous implementation  assumed that there could only be one instance of the provider, this is no longer true. So now those checks are not done and a provider simply created. Another change is that the getter methods for AvailableProperties and HttpHeaders have been refactored for the same reason to use the same logic as PatternWrapper.

In addition to  this, some mistakes were fixed in the Pattern API tests. The Premium and Enterprise namespaces were switched so they have been fixed, and a couple of other typos in the naming have been fixed.

Related Work Items: #1359