TEST: Usage sharing test added, and a check for 20000 User Agents.csv.
A test for usage sharing is carried out by setting up a mock server to
recieve usage sharing data, and another server to send it. Debug usage
sharing options are set via the UsageSharingDebug variable that forces
the data to be sent to the mock server after one request is made. This
then checks that the data sent is gzipped xml with the element 'Device'.

Instead of assuming the User-Agents file exists in "../data", a check is
done. If the module is installed locally from the Device-Detection git
repository it will exist in "../data" (relative to test.js), and if the
module was installed through npm it will need to be placed in the
execution directory. If it does not exist in either location, a message
with a download link is displayed.

Related WOrk Items: #1393 #1083