BUG/FEAT: Changed 'Licence' to 'License' so auto update now works, and now checks for premium or enterprise.
The auto update was failing to return a file properly, this was because
the distributor looks for a 'License' parameter and not 'Licence'. The
spelling has been changed everywhere to be consistent with the change to
the query string which was neccessary fix the update.

The auto updater now checks if the data file being used is premium or
enterprise to get the right download. However, in Trie this is not
contained in the data file, instead the name of the file is used. If the
data file is lite, then an info event is emitted saying that auto update
is not supported.

Related Work Items: #1275