BUILD: Now split into 3 modules.
There are now 3 modules:
fiftyonedegreescore
fiftyonedegreeslitepattern
fiftyonedegreeslitetrie
The first is the main module and the second two just contain the lite
data files and a one line javascript file which exports the path to the
datafile. This means that the path to a lite file can be found with:
var dataFile = require('fiftyonedegreeslitepattern');
which has been applied to all examples. This means that the core now no
longer includes a data file, and the gitCopy.js has been removed from
the fiftyonedegreescore directory so it is not included in the npm
package.

There has also been a change to naming of the modules to only use
lowercase characters, this is the convention on npm.

Changing the directory of the module uncovered some variables which were
set relatively, so this has been fixed.

Related Work Items: #1267 #1213