DOC: Added Windows build steps, test and logging sections.
Added steps exmplaing how Windows installation differs.

Added a logging section explaining how to use the log and the need to
listen for error events.

Added a tests section which shows how to run the mocha tests.

Also updated the readme in general to match the changes made to the API.

Related Work Items: #1271 #1270 #1268 #698