FEAT/TEST: Added cache tests and wrapper functions. Added cache metric function to the Pattern provider so they can be used to test the efficiency of the cache. These functions have also been carried through to the C# wrapper.
Unit tests for the cache have been added. They test performance for different sized caches 100, 1000 and 20000. The logic is first to fill the cache and measure the average detection speed, then fetch all matches from the cache and measure the average speed, then compare the two. These unit tests are in the Cache directory of the unit tests project.
There are a few additions to the unit tests methods to accomodate the new tests:
1. A new CreateWrapper function has been added in PatterBase.cs to make specifying the cache size possible.
2. The DetectLoop functions in Common/Utils.cs have been extended to use a silent parameter which defaults to false. This enables or disables the reporting of detection times and is disabled when doing cache tests and just reported once at the end.
3. A new GetEnumerable function has been added to Common/UserAgentGenerator.cs which takes count as an argument. This uses the GetEnumerable function which takes a count and a pattern, but uses a regex which matches every User-Agent.

Related Work Items: #1374