BUG: Fixed no data file segmentation fault.
When the data file is not declared in the config file, the length of the
filename is not -1 it is 0. so it is now set to the default if length <=
0. In pattern/51Degrees.c, getFileSize is now its own function. It is
called by both getProviderSize functions. Then the returned value is
checked to determine whether to coninue or not.

Related Work Items: #1397 #1400