OPTIM: Added provider id to logging.
A provider id is now set for each new provider and stated when logging.
As a result of this a minor bug was uncovered. Each new provider would
reinitialise the usage sharer when it didn't really need to. Now this is
checked and if the usage sharer already exists an info is logged to say
that an existing usage sharer is being used.

As multiple providers will use the same usage sharer, the provider id is
only stated when it is creating or attaching to the usage sharer.

Related Work Items: #1416