FEAT: Added new V3.2 node.js API. The node.js directory contains three packages: - fiftyonedegreescore - fiftyonedegreeslitepattern - fiftyonedegreeslitetrie The first is the core API, the second two contain Lite data files.
the core is made up of two libraries (FiftyOneDegreesPatternV3 and FiftyOneDegreesTrieV3) which are compiled on install. these are then used by the modules main fiftyonedegreescore. The core also uses update.js to enable auto updating the data files, and shareUsage.js which enables usage sharing.

The usage of the API is detailed in the ReadMe. The structure is the same as the other C based APIs where a provider is created and match objects created from that with the getMatch function. A few additions have been made to make life easier in a node.js package:
- getMatch can be passed an HTTP request object
- Match objects have property getters
- By default boolean properties are converted to boolean types
- Logging is implemented via an event emitter.

The standard API and web examples have been included, and additionaly HTTP and express server examples have been included. All of these are in the node.js/examples directory.

There has been a modification to the 51Degrees.i files. The reason for this is the garbage collector: the time at which the garbage collector will free an object cannot be enforced, so a situation can happen where a match object is not released back to the workset pool and the next time getMatch is called it will hang until the garbage collector runs. To fix this, match objects have no destructor and must be closed manually after use with close() similar to the native C API.

For full usage and examples see https://github.com/51Degrees/Device-Detection/tree/master/node.js