EXAMPLE: Updated Django example to be compatible with 1.10
Changes have been made to the templates section in settings.py as 1.10
now uses a single block for all template options.
The patterns function is now depricated, so urlpatterns is created as a
statndard array type in urls.py.
render_to_response has now been replaced with render in the views.py
which takes a request object and forms the context, the there is no need
to create and pass a context.