DOC: Add git specific install commands.
If "npm install fiftyonedegreescore" were to be executed, it wouldn't
actually install the local version of the package, it wqould install the
most recent from npm. Extra instructions have been added which say to
use a "./" to indicate to npm to use the local directory instead of
looking online for the package.