BUG: Fixed PHP implementation of the find profiles function.
The findProfiles method found in Provider.cpp has two overloads.
However, SWIG does not generate a PHP wrapper for this correctly. To get
arround this, when compiled for PHP, the overload is renamed as a
uniquely named function findProfilesInProfiles. This way
both findProfiles and findProfilesInProfiles correctly return a Profiles
object in the way expected. This has meant a change to the FindProfiles
example which shows the new method of finding profiles within an
collection of profiles.