BUG: Changed value separator to a pipe.
When a comma was used, problems will occur when a value appears which
has a comma in it. For example if the properties for one match were
IsMobile and SupportedBearers, the header could be set to
"True,GPRS,EDGE". A pipe is now used instead, making the header for the
same situation "True|GPRS,EDGE". Readme has also been updated to include
the changes, as well as changing the incorrect 51D_all to 51D_match_all.

Related Work Items: #1691