DOC: Simplified running of PHP examples.
Instead of using apache to demonstrate the PHP examples, a much simpler was is to use the PHP CLI to start a server in the examples directory. This has been updated in the readme. In all examples, the path to the API include has been modified to look from the in place examples directory.