FEAT: Pattern Nginx returns all property values as a list.
Where a list property is specified, the values are now returned as a
pipe separated list the same as Trie. This required the addition of the
fiftyoneDegreesGetPropertyIsList function which returns an integer
indicating whether or not the property can return a list of values. If
the property can return a list, then the maximum value length is
multiplied by a max list length constant (20). This could be calculated
exaclty in the fiftyoneDegreesGetMaxValueLength, but would be
comutationally heavy, so a constant is used.

Related Work Items: #1887