FEAT: Varnish match function now returns all values separated by a pipe.
When a device has multiple values for a single property, all values are
returned as a pipe separated list.
This has meant the addition of an addValue function which is the same as
in nginx, and the buffer size has been increased. The buffer length was
only 24 as this was the most that would ever be used (by device id), now
it is just an arbitrary value which is big enough. This should be
calculated properly like in nginx.

Related Work Items: #1899