BUG: Id property caused problem when set as a required property.
Id is the first property in the strings structure, which means that its index is 0. When the dataset is initialised with this as one of the properties, the requiredPropertyCount was not increased for Id as the validation is ">0", so the requiredPropertyCount was one less than the actual initialised properties. This is now ">=0" so Id is initialised properly like every other property.