FEAT: Added 51D_valueSeparator directive.
The default delimiter to use when concatenating property values is a comma. This can now be changed in the configuration with the 51D_valueSeparator directive. Things to note when setting a separator: 1. Some values of HardwareName contain commas. 2. When getting properties which can return multiple values, they are returned with a pipe separator.

Where a list property is specified, the values are now returned as a pipe separated list the same as Trie. This required the addition of the fiftyoneDegreesGetPropertyIsList function which returns an integer indicating whether or not the property can return a list of values. If the property can return a list, then the maximum value length is multiplied by a max list length constant (20). This could be calculated exactly in the fiftyoneDegreesGetMaxValueLength, but would be computationally heavy, so a constant is used.