FEAT/BUG: Node and Python now share usage data over HTTPS, and fixed usage sharing bug in Python
In Node and Python, usage data is now shared securely over HTTPS. Extra errors ascociated with this change are acounted for in the same way as other web request errors.
In Python, if usage sharing failed, the match function would then fail to return an object as the module would attempt to start usage sharing on a thread that no longer existed. This no longer happens as the stopping flag is checked before attempting to start the thread.