TEST: Optimised tests by running less iterations, and removed usage share test.
Iterating over less User-Agents speeds up the tests and gives the same
result.
The test for usage sharing has been removed in this release, as usage
data is now shared over HTTPS. This causes an SSL error as localhost
does not have a certificate. This may be added back in, but is not
essential for the user.