FEAT: Now uses ws pool and allocates the correct memory for value strings.
The workset pool is now used rather than allocating a new workset for
each request.

The getMaxPropertyValueLength function is now used to allocate the
string which the values are written to. Because of this, the buffer has
been removed and the string written directly to the valueName string
which has been allocated.

Related Work Items: #1938 #2050 #1861