FEAT/CLEANUP: Now takes multtple properties and concatinates values. Also cleaned up functions.
Memory is no longer allocated for the values string, it is all prited to
the reserved memory. The array of properties created also used the
reserved memory.

The printValuesToWorkspace function is now used to loop over all
properties that are requested and print the values directly to the
reserved memory. It returns the length that has been written so a check
can be done. In this way, the getMaxLength does not need to be used as
the memoey does not need to be allocated in the first place. If the
reserved memoey is not enough then nothing is returned and varnish deals
with it.

Related Work Items: #1899 #2019