FEAT: Provider structure is now allocated at module init, and freed when the module is discarded.
The init function allocates just the provider structure, and sets the
free function to be called when discarding the module. The modules
private memory is where the pointer and the free function is set and is
how varnish knows to free it.

Inside the free function, there is first a check to see that the
provider has actually been created by looking at the status. If it has
then it is freed. Then the provider structure is freed no matter what as
this will always be allocated at init.

Related Work Items: #1860 #1861