BUG: Fixed heap corruption by allocating enough memory.
The math to allocate the memory was incorrect as it did not multiply the number of match metrics by the size of the int32_t used to represent its max length. This is now fixed meaning the values are written to memory past that which was allocated.

Related Work Items: #2253