BUG: PHP FindProfiles example thew errors.
FiftyOneDegreesPatternV3.php has now been updated to include the
findProfilesInProfiles method, and the config file edited to update it
on every build in the location used by the examples.
An error in the example has also been corrected. The
findProfilesInProfiles must be used as PHP does not deal well with
overloads, so this change has been reapplied.